refactor: form reusable
