+++ db.json
@@ -575,6 +575,31 @@
       "cv": "https://firebasestorage.googleapis.com/v0/b/react-auth-tutorial-f29d2.appspot.com/o/uploads%2F1%20(8).png?alt=media&token=1cd271f3-1eca-4fa9-b2c1-71397db6aec1",
       "id": "be9f6657-2c32-4016-b186-7aa5fcc0a12e",
       "nama": "Faisal Update Riza"
+    },
+    {
+      "namaPertama": "Dia Update",
+      "namaTerakhir": "Nda",
+      "tanggalLahir": "2024-09-04T17:00:00.000Z",
+      "domisili": "Yogyakarta",
+      "nomorTelepon": "2131233123",
+      "email": "contoh@gmal.com",
+      "tentangDiri": "{\"root\":{\"children\":[{\"children\":[{\"detail\":0,\"format\":0,\"mode\":\"normal\",\"style\":\"\",\"text\":\"dsadsa\",\"type\":\"text\",\"version\":1},{\"detail\":0,\"format\":1,\"mode\":\"normal\",\"style\":\"\",\"text\":\"dasdsa\",\"type\":\"text\",\"version\":1},{\"detail\":0,\"format\":3,\"mode\":\"normal\",\"style\":\"\",\"text\":\"dsad\",\"type\":\"text\",\"version\":1},{\"detail\":0,\"format\":11,\"mode\":\"normal\",\"style\":\"\",\"text\":\"da\",\"type\":\"text\",\"version\":1}],\"direction\":\"ltr\",\"format\":\"\",\"indent\":0,\"type\":\"paragraph\",\"version\":1,\"textFormat\":0,\"textStyle\":\"\"}],\"direction\":\"ltr\",\"format\":\"\",\"indent\":0,\"type\":\"root\",\"version\":1}}",
+      "divisi": "IT Development",
+      "posisi": "Frontend Developer",
+      "kontrak": 2,
+      "gaji": 1000000,
+      "keahlian": [
+        "react"
+      ],
+      "bahasa": [
+        "indonesia"
+      ],
+      "github": "dsads",
+      "linkedin": "a",
+      "status": "idle",
+      "cv": "https://firebasestorage.googleapis.com/v0/b/react-auth-tutorial-f29d2.appspot.com/o/uploads%2FUpdated%20CV.pdf?alt=media&token=7a2e27b2-b3f3-4e89-a962-ada0727bb751",
+      "id": "4f16c8ad-dc02-45c0-a0c4-78aa203695d5",
+      "nama": "Dia Update Nda"
     }
   ]
 }
